requirements.txt
mcp>=1.0.0
pulsar-client>=3.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
typing-extensions>=4.0.0 